requirements.txt
Django==1.4.3
Markdown==2.2.1
South==0.7.6
djangorestframework==2.1.11
gunicorn==0.16.1
psycopg2==2.4.6
wsgiref==0.1.2
